Move plots webview message sending into webview messages
